% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_OOPS_RNase_LFQ.R
\name{raw_OOPS_RNase_LFQ_data}
\alias{raw_OOPS_RNase_LFQ_data}
\title{PD output for LFQ of RNase +/- OOPS experiment}
\description{
F17-20 = RNase(-) reps 1-4
}
\details{
F21-24 = RNase(+) reps 1-4
}
\section{OOPS_RNase_LFQ/LFQ_OOPS_RNase_PeptideGroups.txt}{


This is the peptide-level output from PD
}

\section{OOPS_RNase_LFQ/LFQ_OOPS_RNase_PSMs.txt}{


This is the PSM-level output from PD
}

\section{OOPS_RNase_LFQ/LFQ_OOPS_InputFiles.txt}{


This is the InputFiles output from PD
}

