% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_OOPS_SILAC_data.R
\name{raw_OOPS_SILAC_data}
\alias{raw_OOPS_SILAC_data}
\title{PD output for OOPS CL vs NC using SILAC labelled U-2 OS cells}
\description{
For replicates 1 & 2: H=CL, L=NC
}
\details{
For replicates 3 & 4: H=CL, L=NC
}
\section{OOPS_SILAC/OOPS_*_PSMs.txt}{


These are PSM-level outputs from PD
}

\section{OOPS_SILAC/OOPS_*_PeptideGroups.txt}{


These are Peptide-level outputs from PD
}

